CC/requirements.txt
flask
flask_cors
gunicorn
mysql-connector-python
python-dotenv
numpy
Pillow
google-cloud-storage
tensorflow
PyJWT
bcrypt
pytz